# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Face Swap API
  description: |
    Backend for the iOS Face Swap app. Uses Novita.ai for face merge, video face swap, and motion-controlled video generation.

    **Key points for iOS:**
    - All request bodies are **JSON with base64** (no multipart). Encode images/videos as base64 strings.
    - **Sync:** `POST /api/v1/merge-face` returns the result image in the response.
    - **Direct result:** `POST /api/v1/video-merge-face` and `POST /api/v1/motion-control` submit the job to Novita and return the task result directly in the response.
    - Base URL: your Railway deployment (e.g. `https://your-app.up.railway.app`).
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: /
    description: Current host (use your Railway URL, e.g. https://your-app.up.railway.app)
...
paths:
...
  /api/v1/users/me/credits:
    get:
      tags: [Users]
      summary: Get user credit balance
      parameters:
        - name: X-Device-ID
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Credit balance
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    type: integer
  /api/v1/webhooks/revenuecat:
    post:
      tags: [Webhooks]
      summary: RevenueCat webhook
      description: |
        Called by RevenueCat for subscription events.
        Requires Authorization: Bearer <REVENUECAT_WEBHOOK_SECRET>.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Event received
...
  /api/v1/categories/{category_id}/templates:
    get:
      tags: [Templates]
      summary: List templates in category
      description: Returns all templates for the given category, ordered by created_at ascending.
      operationId: listTemplates
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of templates
          content:
            application/json:
              schema:
                type: object
                properties:
                  templates:
                    type: array
                    items:
                      $ref: '#/components/schemas/Template'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [Templates]
      summary: Create template in category
      description: Creates a new template in the given category.
      operationId: createTemplate
      parameters:
        - name: category_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [novita_model_id]
              properties:
                novita_model_id:
                  type: string
                preview_url:
                  type: string
                  nullable: true
                credit_cost:
                  type: integer
                  default: 0
      responses:
        '201':
          description: Template created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/templates/{template_id}:
    get:
      tags: [Templates]
      summary: Get template
      description: Returns a single template by ID.
      operationId: getTemplate
      parameters:
        - name: template_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags: [Templates]
      summary: Update template
      description: Updates an existing template. All fields are optional.
      operationId: updateTemplate
      parameters:
        - name: template_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                novita_model_id:
                  type: string
                preview_url:
                  type: string
                  nullable: true
                credit_cost:
                  type: integer
      responses:
        '200':
          description: Updated template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Templates]
      summary: Delete template
      description: Deletes a template by ID.
      operationId: deleteTemplate
      parameters:
        - name: template_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/users:
    post:
      tags: [Users]
      summary: Get or create user by device_id
      description: |
        Pass `X-Device-ID` in the body. If a user exists for that device_id, returns 200 with the user.
        If not, creates a new user and returns 201.
      operationId: getOrCreateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [X-Device-ID]
              properties:
                X-Device-ID:
                  type: string
                  description: Unique device identifier (e.g. UUID from iOS)
                onboarding_answers:
                  type: object
                  description: Optional onboarding answers (only used on first creation)
      responses:
        '200':
          description: Existing user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '201':
          description: New user created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Missing X-Device-ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
...
    Template:
      type: object
      properties:
        id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        novita_model_id:
          type: string
        preview_url:
          type: string
          nullable: true
        credit_cost:
          type: integer
        created_at:
          type: string
          format: date-time
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        device_id:
          type: string
        onboarding_answers:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time